Ticket #1189
Add the xml:lang attribute to LangMarks plugin.

The configuration for this plugin has changed, old configurations are accepted with a warning to the console.

By default BOTH lang="" and xml:lang="" are emitted for compatability, simply remove one or the other from the LangMarks.attributes array.

{{{
Xinha.Config.prototype.LangMarks = {
  'languages': [
    { name:"Greek",   code: "el" } ,
    { name:"English", code: "en" },
    { name:"French",  code: "fr" } ,
    { name:"Latin" ,  code: "la" }  
  ],
  
  'attributes': [
    'lang',
    'xml:lang'    
  ],
};
}}}