#1226 - Gecko problems with paraHandlerBest
I actually checked the existing paraHandlerBest in FF 58 to see if I could reproduce the problem still and I couldn't so maybe FF fixed it anyway in the 7 years since it was first reported. 

But despite that I'm committing the paraHandlerBest.js contributed by ejucovy and douglas, since it seems to work just as well and they probably actually use Firefox, so, yeah.