#1480 - Fix some InsertAnchor oddities
InsertAnchor now does not preserve the selection (wrap the anchor around it) when inserting, this is because this made dragging anchors weird.  It's better to have it just insert an atomic anchor.

Also remove the background image and just use the placeholder image, this fixes problem with bad CSS getting dumped into the anchor when it's dragged.