#1476 Improve ContextMenu behaviour for misspelt words (detected by browser).
If a words is misspelt according to the browser (red squiggly line inserted by browser) and you select that word and only that word (in Chrome this will automatically happen when you right click it, in Firefox double click it) then right click the ContextMenu plugin will fall back to the browser's context menu and allow you to choose spelling suggestions.

If you right click within a link, or on a selected image, or multiple words, or without highlighting the word (just caret, which for misspelt words on chrome may not be possible) then the ContextMenu will open as normal.