Add a new plugin, WebKitResize, which adds drag-resizing capability to images tables and cells in WebKit browsers (and does nothing in non-webkit browsers).
Uses jQuery, if you load your own jQuery version before registering the plugin (configuring Xinha) then it will be used, otherwise it will load a version of jQuery of it's own choosing (currently 3.3.1, but it wil work at least back to 1.7 I believe).

It is strongly recommended that your page (containing the editor, not what you are editing) is in STANDARDS MODE, that means preferably using a doctype such as 

  <!DOCTYPE html>

Quirks mode (no doctype at all is the easiest to remember) also works OK, "Almost Standards Mode" not so much.

Tested in Chrome only at this stage.

Code is from

  http://www.editorboost.net/Webkitresize

and is GPLv3 licenced.