Add FancySelects plugin to resolve #1365, along the way move jQuery and MooTools into libraries/ improve AssetLoader and add Xinha.loadLibrary.
The `loadScript` and `loadStyle` functions of Asset Loaders have been changed to load the given script/style only once, this was the most common situation.  `loadScriptAlways` and `loadStyleAlways` have been added with which you can force one to be reloaded.

loadLibrary presently just accepts a library name, `jquery` or `mootools` and loads the latest version in the libraries folder (hard coded), eventually we may need to have some sort of version-requesting, but cross that bridge if we come to it.