Improve the undo/redo handling considerably by preserving the caret position.
That is, each snapshot saves where the caret was, and undoing to a snapshot puts the caret back in that position.

So if you are typing along and hit ctrl-z to undo what you wrote, you don't get dumped back at the start (or end depending on browser) of the document but can carry on typing immediately where you left off.

Discussion in #360