By default, load a merged language file if it exists for the given language in `lang/merged/*.js`
Merged files are created using the `contrib/lc_create_merged_file.php` script and combine all the language files into one, mainly for easier translation work, but also can be used handily for this reduction in browse requests.

Merged files should be included in the distribution by default.

Note that when a merged file is in use, any other existing language file for a context included in the merged file, is not consulted.  Hence the recommendation in the comments of each language file that people normally should not edit that (it would be recreated from lc_split_merged_file.php when the submit their merged file back for inclusion in the distribution).