Change the initial iframe src to `javascript:''` and change the iframe initialisation so it's not reliant on the load event.
This is to allow for using Xinha from an external server, that is, a CDN, previously same-origin restrictions on the iframe prevented this from working because the iframe was provided with the "blank.html" file to load, which would come from the CDN and so Xinha having by then be loaded into the containing document on a different origin did not have permission to edit the document.

I've tested in a variety of browsers and `javascript:''` appears to be pretty reliably working.

Credit does have to go to TinyMCE's source code which appears to use `javascript:''` in some cases, and that's where I got the idea from after hunting to see how they solved the problem.

Of couse, you can use Xinha from a CDN now, but not the PHP based plugins, namely MootoolsFileManager and Linker, of course these plugins need to run on the server to which you are uploading/selecting files and links, so they can't work in a CDN environment.  It would be possible I expect to direct Linker to a scan.php implementation on your server, but MootoolsFileManager is a bit much to ask.  POSSIBLY an `external_plugins` could be made, but then you are dealing with things which are outside of `_editor_url`.