Add `xinha_config.resizableEditor = [bool];` option to allow drag-resize of editor.
This works pretty well in Chrome.  Not at all in IE/Edge or Firefox, and works to an extent in Safari.  

The differences are due to how much support there is for the css `resize` property, specifically, Firefox doesn't support it on Iframe properly, IE/Edge doesn't support it at all, and Safari's support is a bit weird especially in 9.x, in 11.x it only allows you to increase size, not decrease.