Some slight toolbar modifications and improve resizability.
Add some linebreaks to the toolbar between the standard drop-down selectors, this allows the toolbar to flow to smaller widths.  

Also fix a resizing problem with the editor due to the status bar pushing it out when there is a lot of path information being shown.  The statusBar needs to be looked at to make this better when there is too much path information to show, see #1728